+++ constants.py
@@ -4,11 +4,14 @@ ARTICLE_TEXT_END_MARKER = "</div>"
 ARTICLE_TEXT_UNICODE_MARKER1 = "\u200e"
 ARTICLE_TEXT_UNICODE_MARKER2 = "\u200f"
 ARTICLE_DATE_MARKER = "post-date-value"
+ARTICLE_PREFIX_CONTENT = "SPONSR_CONTENT "
 ARTICLE_PREFIX_DATE = "SPONSR_DATE "
+ARTICLE_PREFIX_TITLE = "SPONSR_TITLE "
 ARTICLE_PREFIX_NEXT = "SPONSR_NEXT "
 ARTICLE_PREFIX_SRC = "SPONSR_SRC "
 ARTICLE_PREFIX_URL = "SPONSR_URL "
 CSS_PICKER = "div[class^='datepicker--cells datepicker--cells-days']"
+FILE_CACHE_CONTENT = "/tmp/sponsr-content"
 FILE_CACHE_TOC = "/tmp/sponsr-toc"
 FILE_CACHE_VISIT = "/tmp/sponsr-visit"
 FIRST_POST = "https://sponsr.ru/marahovsky/?post_date=1585688400000"
