+++ run.py
@@ -1,12 +1,12 @@
-from DoGenerateContent import *
+from DoExtractContent import *
 from DoGenerateTOC import *
 from DoVisitArticles import *
 from constants import *
 
 tasks = [
     #DoVisitArticles(FILE_CACHE_VISIT),
+    DoExtractContent(FILE_CACHE_CONTENT, FILE_CACHE_VISIT),
     DoGenerateTOC(FILE_CACHE_TOC, FILE_CACHE_VISIT),
-    DoGenerateContent(FILE_CACHE_VISIT),
 ]
 
 for task in tasks:
